// src/frontend/src/api/models/TitleCreate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TitleCreate = {
    /**
     * Title name
     */
    name: string;
    /**
     * Edition number
     */
    edition: number;
    /**
     * Newline-separated list of authors
     */
    authors: string;
    /**
     * Year of publication
     */
    yearOfPublication: number;
    /**
     * Newline-separated list of tags
     */
    tags: string;
};